comparacion fechas en php

+++ TP/server/app/Http/Controllers/PostulacionController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Postulacion;
+use App\Models\Llamado;
 use Illuminate\Support\Facades\DB;
 use Exception;
 
@@ -24,17 +25,27 @@ class PostulacionController extends Controller
   {
     if ($request->id_llamado && $request->curriculum_vitae) {
       if(!Postulacion::where([['id_persona', auth()->user()->id], ['id_llamado', $request->id_llamado]])->first()) {
-        try {
-          $postulacion = new Postulacion();
+        $llamado = Llamado::find($request->id_llamado);
+        if ($llamado) {
+          $fechaDeHoy = strtotime(date('Y-m-d'));
+          if ((strtotime($llamado->fecha_fin) >= $fechaDeHoy) && (strtotime($llamado->fecha_inicio) <= $fechaDeHoy)) {
+            try {
+              $postulacion = new Postulacion();
 
-          $postulacion->id_persona = auth()->user()->id;
-          $postulacion->id_llamado = $request->id_llamado;
-          $postulacion->estado = "Postulado";
-          $postulacion->curriculum_vitae = $request->curriculum_vitae;
-          
-          $postulacion->save();
-        } catch (Exception $e) {
-          return response()->json(['error' => $e->getMessage()], 406, []);
+              $postulacion->id_persona = auth()->user()->id;
+              $postulacion->id_llamado = $request->id_llamado;
+              $postulacion->estado = "Postulado";
+              $postulacion->curriculum_vitae = $request->curriculum_vitae;
+              
+              $postulacion->save();
+            } catch (Exception $e) {
+              return response()->json(['error' => $e->getMessage()], 406, []);
+            }
+          } else {
+            return response()->json(['error' => 'No se puede inscibir en esta fecha'], 406, []);
+          }
+        } else {
+          return response()->json(['error' => 'No existe el llamado'], 406, []);
         }
       } else {
         return response()->json(['error' => 'El usuario ya se encuentra postulado al llamado'], 406, []);
@@ -47,14 +58,24 @@ class PostulacionController extends Controller
   public function eliminarPostulacionDelUsuario($id_llamado)
   {
     if ($id_llamado) {
-      try {
-        $postulacion = Postulacion::where([['id_persona', auth()->user()->id], ['id_llamado', $id_llamado]])->first();
+      $llamado = Llamado::find($id_llamado);
+      if ($llamado) {
+        $fechaDeHoy = strtotime(date('Y-m-d'));
+        if ((strtotime($llamado->fecha_fin) >= $fechaDeHoy) && (strtotime($llamado->fecha_inicio) <= $fechaDeHoy)) {
+          try {
+            $postulacion = Postulacion::where([['id_persona', auth()->user()->id], ['id_llamado', $id_llamado]])->first();
 
-        if ($postulacion) {
-          $postulacion->delete();
+            if ($postulacion) {
+              $postulacion->delete();
+            }
+          } catch (Exception $e) {
+            return response()->json(['error' => $e->getMessage()], 406, []);
+          }
+        } else {
+          return response()->json(['error' => 'No se puede dar de baja en esta fecha'], 406, []);
         }
-      } catch (Exception $e) {
-        return response()->json(['error' => $e->getMessage()], 406, []);
+      } else {
+        return response()->json(['error' => 'No existe el llamado'], 406, []);
       }
     } else {
       return response()->json(['error' => 'Ingrese el id del llamado a eliminar'], 406, []);
